move path functions to pkg package
