+++ io/pkg/path_test.go
@@ -1,4 +1,4 @@
-package content
+package pkg
 
 import (
 	"testing"
