+++ workflow/engine/jsonnet_template_engine.go
@@ -35,6 +35,9 @@ func NewJsonnetTemplateEngine(fs *afero.Afero, context *config.GFlowsContext, co
 
 func (engine *JsonnetTemplateEngine) GetWorkflowSources() []string {
 	files := []string{}
+	// TODO: replace GetWorkflowDirs() with GetLibs() and add .WorkflowDir() and .LibDir() to each lib.
+	// ..altho: how to deal with local dir?
+	// Perhaps add GFlowsPackage interface or something to abstract over libs and context.
 	for _, workflowsDir := range engine.env.GetWorkflowDirs() {
 		err := engine.fs.Walk(workflowsDir, func(path string, f os.FileInfo, err error) error {
 			ext := filepath.Ext(path)
